距离判定要以米为单位计算 以厘米为单位容易因为int值超限导致结果错误

+++ Creature3DApi/include/crnetapp/appNetMethod.h
@@ -2326,7 +2326,7 @@ protected:
 	crInstanceItem* m_this;
 	CRCore::ref_ptr<crInstanceItem> m_item;
 	//_crInt64 m_distEnmity;
-	int m_dist;
+	float m_dist;
 };
 class crCalcEnmityByDistRateMethod : public CRCore::crMethod
 {//SceneServer Method
